fix(llma): extract cost model matching
